Remove references to User->getData() (#7334)
The getData function is a code smell which we still have scattered throughout our code. It loses type information and generally makes it harder for static analysis tools (or humans) to figure out what our code is doing.

This updates most of the references to use the User class's getters for the appropriate field, which does not lose type information.

There are a couple exceptions which were not updated: 1. 2 places in the code use getData() to send the all the user's data to a template (my_preferences and edit_user). 2. Unit tests for the user class use it to get test the internal state. These situations need to be fixed/tested separately before the function can be completely removed.