Add interface to replace method_exists('getCenterID') checks (#7337)
This adds 2 interfaces, SiteHaver and MultiSiteHaver to implement on classes that implement getCenterID and getCenterIDs functions respectively. This replaces the lossy method_exists checks for the 2 methods, which only verifies that a method with that name exists, and doesn't validate any type information such as arguments or return types. (In fact, it turns out there's a class in the examiners module which has an unrelated getCenterID($sitename) function, it's just luckily never passed to a pass which does a method_exists check).

This has a few benefits:

    Static analysis tools are better able to analyze types
    Some code can be simplified, since the types/nullability are explicit in the interface they don't need to be verified inside the method_exists checks
    Some hacks that were necessary because of the method_exists checks (such as assigning arguments to a temporary variable and coercing the types for phan to recognize the methods exist inside of the method_exists blocks) aren't necessary inside of instanceof checks.
    If we change the signature (or decide other methods should be added) it can be updated in the interface to ensure we don't miss any classes that were implicitly implementing the getCenterID contract.