Remove most references to Candidate->getData (#7335)
This is analagous to #7334 but for Candidate->getData. It removes calls to getData in the candidate class and replaces them with appropriate getters. There are a few places where the array for getData is passed to a smarty template and some unit tests which rely on it, and these are untouched as they need more thorough testing

This should improve the ability of phan to track our type information while it analyses the code.