[login] module upgraded for react (#5828)
Upgraded module to use react. I had to fix some things in LORIS to get this all to work and there are multiple changes in some JSX files as well.

What was updated to react:
Login, password expired, request account, and reset password.

The Content-Security policy was broken for google captcha and this fix includes the iframe that recaptcha uses. 

    Redmine #16050
    Resolves #5828
    Resolves #5214