[issue_tracker] Only show inactive user if already selected in form (#7306)
Currently, the issue tracker shows inactive users. You are able to create a new issue and assign it to them.

This filters out inactive users so that they don't show up in the list of assignees in the 'New Issue' form. In the 'Edit Issue' form, the inactive user appears in the list of assignees but the option is disabled. this is so that inactive users that are already assigned show up but cannot be reselected.