New Data Dictionary framework and module (#6936)
This updates the LORIS Data Framework to include a concept of a data dictionary that a module may manage. It adds a new "dictionary" module to manage the module dictionary but leaves the datadict module in place because it's currently the only place to manage the parameter_type table used by the DQT.

The extension of the data framework to include dictionaries also includes many enhancements from the current "data dictionary" in LORIS which only includes a field name and description. Types, Scopes, and Cardinalities are added which more thoroughly describe the data and will eventually be able to be used to build richer data-type sensitive interfaces in places like the DQT. DictionaryItems implement the AccessibleResource interface to enable modules to give column-level granularity over data.

Modules provide categories of DictionaryItems, which allows them to group items in the frontend. This is primarily to make it possible to filter the instruments from the "instruments" module (in the current data_dict module, the Source From options are literally populated by a call to \Utility::getAllInstruments()) in a manageable way, but can be used by any module that wants to group the fields.

Instruments written in PHP must now implement a new getDataDictionary() abstract function to describe the fields provided by the instrument, but a LorisFormDictionaryImpl is provided which populates it in the same way that the lorisform_parser uses it to populate the parameter_type table.