[Core] Update NDB_Page constructor to include LorisInstance object (#7414)
This updates the NDB_Page constructor to include a LorisInstance object. The LorisInstance object represents the properties of the LorisInstance itself (such as the database connection) and can replace many factory/singleton calls, but is seldom used in LORIS because it's difficult to get a copy of the object. (It needs to be indirectly accessed from a PSR7 request attribute). After moving it to the constructor, it can now be accessed directly from any NDB_Page class or subclass in a simple class property.