[Instruments] Fix Display of 0 scores with json data (#7443)
When a JSON data instrument has a score column that's saved
as 0 (the int value, not the string), the LORIS front end
does not display it, and instead displays the field as a header.

This is caused by PHP's weak equality treating `"string" == 0` as
true. Forcing the template to use a strict equality comparison
fixes this without affecting the header columns that the comparison
in the template was intended to catch.

Fixes #7441.