AnonymousUser getters override to prevent fatals (#7225)
User::factory() allows a nullable username argument. If null, the factory returns an instance of AnonymousUser and let userInfo unset. As most of User's getters rely on userInfo to return a value, such AnonymousUser's getter will return a null value, which contradicts their non-nullable type and throws an error.

This overrides the AnonymousUser's getters to return valid values..

    Resolves #7223