[TimePoint] Add TimepointData DTO class (#7432)
Add a TimePointData class to act as a data transfer object for
data related to TimePoints.

A TimePointData object is a DTO (data transfer object) to
store a representation of data associated with a TimePoint
object in LORIS.

The TimePointData may be partially loaded, in which case any
getter trying to access an unknown will throw an exception.
This allows modules to only load relevant data in a query,
and avoid the overhead of the TimePoint::singleton() constructor.

For instance, a query that selected CenterID and ProjectID on
the database could instantiate a TimePoint as:

$visit = new TimePoint(new TimePointData(null, $projectid, $centerid));

with data loaded from the query, and then call

$visit->isAccessibleBy($user) without having to load all the
over data that TimePoint::select() does. If an attempt is
made to access a property not in the DTO (session in this example),
an exception is thrown, rather than silently returning the incorrect
value.