Instrument class and LINST class cleanup (#7183)
Deprecation of begintable and endtable in LINST instruments. these elements do not appear to be used anywhere and when tested they only seemed to corrupt the HTML without bringing any advantages. The instrument template automatically converts groups to tables (behaviour that should be changed in the future) so removing these elements will just bring LINST to par with PHP instruments for now

Added a bunch of PHP docs on variables not having them

Deletion of the dateTimeFields and monthYearFields class arrays. Projects still maintain off and on these arrays which are unused anywhere. These arrays are being populated but their usage died with the LorisForm replacing quickform. Here is an example of where they were used

Deletion of the unused _getDatabaseDate() function