Add PSR3 Logger to LORIS
This adds a PSR3 logger implementation to LORIS. This PR implements a single
logger type, ErrorLogLogger, which logs to the web server's standard error_log
based on the current log level verbosity configuration in the LORIS config
module.

Three types of logs are implemented in this PR, the verbosity: 1. Database logging
2. HTTP Request logging and 3. exception logging