[DQT] Check for scan types in MRI Importer script
Add a check to verify that GetScanTypes returned some
valid, configured scan types with data in CouchDB import
script.

If the import script is run with no scan types, it currently
only imports the session QC status. The fact that the script
failed to import the important data isn't caught until after
a user tries to use the DQT.

This adds a check to error out if getScanTypes does not return
any scan types at import time.