Add Alias in parameter type table and map BIDS Alias to MINC headers present in parameter_type (#7503)
A new column is added to the parameter_type to add an Alias to the parameters. For imaging related fields that comes from the MINC header of images, this Alias will be used to store the BIDS term mapping for a given parameter. This will be needed for the dcm2bids LORIS insertion pipeline in order to map the MINC terms with the BIDS terms.

It should be noted that currently, some fields are present multiple times in the parameter_type table with different parameter name (this can be observed in the raisin bread dataset). An example would be the echo time. There are 3 different parameter types available for the echo time: echo_time (ID 130), acquisition:echo_time (ID 113) and dicom_0x0018:el_0x0081 (ID 209). For the BIDS terms, only the echo_time has the BIDS term Alias EchoTime to avoid mapping EchoTime to 3 different parameter type ID.