[user_accounts] Use userID to update examiners table (#7462)
Currently when updating the examiners table in edit_user, it is assumed that a unique examiner exists based on their full name instead of their userID. This causes 2 problems:

- If an examiner with the same name as another examiner is edited, the older one will be replaced in the table and will no longer be considered an examiner. I.e. only one examiner can exist per full name.
- If a user's name is edited, then a new row will be created in the examiners table and there will be 2 rows for the same user.

This fixes this issue by assuming in the user_accounts module that examiners have a unique userID instead of assuming examiners have a unique full name in the user accounts module.

However, since examiners do not have to be users, the examiners module was edited to assume that an examiner added is not a user. The examiner module does not ask for any identifying info other than the full name, so it would be impossible to know which user is being referred to if we are accepting that different users may exist with the same name.

The resulting implementation is that only one non-user examiner can be added per full name, but any LORIS users that are added as examiners do not have to have a unique name.

The instrument library file is also edited to add an email beside the full name in the examiners list examiners with duplicate names.

Included is a cleanup script that gets rid of the duplicate userIDs in the examiner table which were added because of a name change in "edit user".

A script was not made for adding back examiners that may have been overwritten by a new user with the same name, because there is no way to determine this based on the previous implementation of examiners.