[Visits Architecture (2 of 4)] - Making config.xml unnecessary (#7729)
This removes the need for the visits-subprojects association to be defined in config.xml. as we are trying to slowly phase out config.xml and since the database already has all of the data in various `_rel` tables.

All references to getSetting("VisitLabel") were removed and replaced by the appropriate and equivalent code.