[conflict_resolver] Fix Correct Answer filter on Resolved Conflicts page (#7727)
This fixes the Correct Answer filter on the Resolved Conflicts page. Previously, the page was displaying the correct answer but filtering by the value in the database which can only be 1 or 2 depending on which of the 2 conflicting values was selected. This is solved by updating the provisioner to return the correct answer value instead of the 1 or 2.

    Resolves #7702