[bvl_feedback] Remove redundant bvlFeedbackPossible method
The middleware makes a call to NDB_BVL_Feedback::bvlFeedbackPossible
to determine whether the feedback icon should be displayed. This
function checks the test_name (based on the page passed by the middleware) and user permission to filter out who should not see the
icon. However, permission is already checked by the middleware, and
the test_name check breaks the loading of the icon for instruments
since the page passed by the middleware is the NDB_BVL_Instrument
(sub)class, not the "instruments" module.

The middleware checks for the existence of a "getFeedbackPanel"
method on the page (which is more robust, and exists on the instrument class),
so the test name check is unnecessary and only breaks instrument feedback.
The result is that the template variable can just be set to "true" if we're in
the if statement without loss of functionality, and the static method
(which is unused anywhere else) can be removed.

Resolves #7665