[test] Add safeTextContent method and use it
One of my PRs randomly failed on mri_violations tests. The
problem appears to be that code is using javascript->execute
script to get the text content of some elements. However, if
an action was just triggered by javascript, the element may
not yet have appeared on the page. This

1. Extracts the wait code from safeFindElement to a waitForElement
   wrapper, to reduce boiler plate when a test needs to wait
   for an element
2. Adds a safeTextContent helper, which will wait for an element
   to appear before running executescript on it.
3. Updates various tests (ones that I found by either grepping the
   code for "sleep" or were randomly failing on other PRs) to use
   these methods.
4. Changes some execute scripts which just do a "click" to use
   the analogous safeClick method (which already exists).

This is not done throughout all tests yet, but hopefully fixes some
of the problematic ones.