[survey_accounts] Fix issue where multiselect value is of type string instead of array when only one value is selected (#7739)
The function in_array(string needle, array haystack) expects the parameter haystack to be of type array. When mapping the multiselect values from the DB, we expected the special chars {@} to be present in the string and only exploded the string if it was. In the case when only one value is selected the chars {@} aren't present and so the string is never split into an array. This would then cause the haystack value to be of type string instead of array. This PR removes the check for the special chars {@} and just always calls the explode method no matter if the {@} is present or not. by doing this it will always force the string to be converted to an array.

    Resolves  #7675