[brainbrowser] Remove unused file
There do not appear to be any references to MincEnv.php.inc in the
code. It doesn't have any classes defined, so it's not autoloaded.

I believe this file dates to the time before BrainBrowser could load
MINC files and depended on the minc toolkit, which is no longer the
case.