[BrainBrowser] Fix duplicate download of image data
This fixes an issue where image data is downloaded twice by
BrainBrowser. The brainbrowser.loris.js file makes a request to
image.php to get metadata for the file from http headers, and then passes
the URL to brainbrowser which makes a separate data.

Replace the GET request with a HEAD request from the LORIS side so
that only brainbrowser downloads the data.

Resolves #7805