[battery_manager] Fix getCenterIDs to return CenterIDs
This updates the battery_manager Test class to correctly
return the type that it claims to return, as required by
the MultiSiteHaver that it claims to implement. This type
incompatibility was not caught by phan because the docblocks
were using a non-specific 'array' type instead of the more
accurate '\CenterID[]' that the interface requires.

Resolves #7798