[Imaging Browser] Sequence Type filter fix (#7565)
Change the keys of the multiselect to make them match the values in the cells. It use to be [scan_type_id => scan_type_label] because the scan_type_id was sent to the backend to apply the filters. Now that the filtering is done in the frontend, the options needs to be [scan_type_label => scan_type_label].

    Resolves #7563