[brainbrowser] Remove ajax from module
This updates brainbrowser in LORIS to get rid of the ajax directory.
In doing so, it also updates the architecture to be somewhat more
sane.

Previously, a request would be made from javascript to an ajax script
which takes a FileID and returns the raw file content with some metadata
in HTTP headers. This request would be made once per file to be displayed in BrainBrowser. The body of the request would then be discarded, and the
same endpoint would be passed to brainbrowser. Brainbrowser would download
the file content a second time, and after all files were re-downloaded
display the volume viewer.

In the new architecture, an imageinfo endpoint is added which returns
the metadata in JSON format. Included in the information in the response
is a URL which can be passed to brainbrowser to download the raw file
data. The LORIS API is used for imaging browser files, and a new endpoint
is added to the mri_violations module for files which can not be accessed
through the API.

The end result is that the ajax directory is no longer used by brainbrowser,
the API is used to download files, and some more access checking is
done on accessing mri violations files.