[DQT] Fix broken file link when querying using 'Add All' (#7823)
Downloadable files i.e. minc files are missing their download links when the 'Add All' button is used to select a query.
This fixes that by fixing a typo found.
This also fixes a typo in the CouchDB MRI import script that fails to find any scan types to import.