[Config] Remove unused extLibs setting from DB
This removes the extLibs path from the database. The config
setting was at one point used to locate the system version
of smarty, but this has not been the case since composer was
used to manage external libraries in January 2015.

Now it's just an arbitrary setting in the config module that
with the value "PATH/TO/EXTERNAL/LIBRARY/" which has no impact
other than adding "PATH/TO/EXTERNAL/LIBRARY/" to the include
path (which is also unused, since classes are autoloaded.)