[dqt] New ProgressBar/Remove Smarty/with needed fixes for performance (#7516)
This contains the enhancements that had been done for the ABCD project. The smarty template had to be removed to add the ProgressBar and a few places in the DQT had to be modified to load without any sessions or visits data. Note this change helps when the couchdb has lots of sessions for the project.