Fix type of physiological electrode X, Y and Z fields to be nullable (#7849)
The values stored in the X, Y and Z columns of the `*electrodes.tsv` BIDS file can be set to n/a which cannot be inserted into the current type of the corresponding fields in the physiological_electrode table. This modifies the default schema and adds a patch to make those columns NULLABLE so the values of X, Y and Z can be set to NULL when `n/a` is provided in the BIDS `*electrodes.tsv` file.