[LINST] Fix standard dates always required bug (#7925)
XIN rules always required Standard Date elements to be completed, even when user tries to override rules from the .rules file. The bug is caused by the stripping of the _date suffix for datefields (the suffix is later added in the Instrument class to make LINST and PHP instruments compatible) which means the incorrect question name was being passed to validation without the _date suffix.

This moves the addition of the question to the validated array to BEFORE the _date is stripped.