[Battery] Fix error when multiple rows of the same instrument match (#7927)
2 Issues are fixed By the following changes, both related to the fact that when multiple entries exist in the test_battery that have different attributes (not duplicates) but all match the lookup at a specific timepoint the next_stage and instrument_list modules fail.

Issue 1: The next stage module calls on the battery class to populate the flag table when the visit is started. When multiple entries of the same instrument are returned from the lookup query, the battery class attempts to create both which triggers a duplicate error in MySQL

Issue 2: The instrument_list module uses the battery class' getBatteryVerbose function to list the instruments properly grouped and ordered. the issue occurs when multiple entries of the same instrument are found in the lookup results. This PR removes the duplication of the instrument in the front end when that happens

Resolves #7130