[imaging_browser] Remove jQuery from imaging browser (#7833)
This removes the couple references to jQuery in the imaging_browser
so that the errors:

```
/home/runner/work/Loris/Loris/modules/imaging_browser/jsx/ImagePanel.js
26:5  warning  The jQuery constructor is not allowed  no-jquery/no-jquery-constructor
26:5  warning  .tooltip is not allowed                no-jquery/no-other-methods
126:5  warning  The jQuery constructor is not allowed  no-jquery/no-jquery-constructor
126:5  warning  .DynamicTable is not allowed           no-jquery/no-other-methods
134:5  warning  The jQuery constructor is not allowed  no-jquery/no-jquery-constructor
134:5  warning  .DynamicTable is not allowed           no-jquery/no-other-methods
```

are not displayed. The first error was a call to ".tooltip" which was
unnecessary. The value of the tooltip is the same as the value of the
panel header being hovered over.

The second was a call to .DyanmicTable on the panels. This is cosmetic
and should have little effect (I don't think I've ever seen the DynamicTable
rendered in the imaging browser.)

The third call was just undoing the second call to prevent errors when
the page is unmounted.