Add S3 Support to LORIS (#7963)
This adds support for accessing files via S3 to LORIS. The AWS SDK is used so that s3:// URLs can be accessed as normal files within PHP if the environment variables AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY are set (which roughly correspond to AWS username and passport). One of the LORIS config settings AWS_S3_Region or AWS_S3_Endpoint must be set for the SDK to work (the latter for AWS hosted instances, and the latter for self-hosted.)

The LORIS API for imaging is updated (for both thumbnails and the mnc files). The imaging_browser and brainbrowser have already been updated to use the LORIS API, so should work transparently.

Other modules that deal with files (document_repository, media, issue_tracker, etc) will need to be tested separately. (The most likely cause of bugs is that a path config option is concatenated to the filename at some point, invalidating the url.)