[Login] Request account with bad email UI seems to work (#7908)
Updates the email component to use the pattern property for the email input. Also using the type property for the message to displayed to the user when email isn't matching the pattern property. See here: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/email

edit: I added frontend logic for if by any chance the php validation finds an error not caught by the frontend and so the user will get a notification that email is invalid.

    Resolves  #7897