Move DB credentials to environment
This prevents the possibility of PHP printing a stack trace
with database credentials if an exception is thrown from the
database constructor, by moving the credentials from function
arguments to environment variables.

Since the function signature changed, all calls to \Database::singleton
were (finally) updated to use LorisInstance->getDatabaseConnection (if
available) or NDB_Factory->database (if no LorisInstance available in
the current context) in order to centralize the change to one place.

Resolves #8014