Add PHP 8.1 Support (#7989)
This fixes the loading of the main LORIS pages with PHP 8.1.

Between 8.0 and 8.1 the behaviour of the PDO changed so that database selects are typed according to the column, while previously they were always a string. This restores the behaviour of 8.0 by always casting them with strval to avoid fatal type errors.

Some signatures for jsonSerialize also had to be modified to prevent warnings and phpstan had to be upgraded.