[Readthedocs] Updating requirements file to fix RTD build (#8120)
A new version of the "jinja" package is causing the Read the Docs build to fail, so the Read the Docs page cannot be updated to the latest version of the documentation. This PR adds a line to the readthedocs/requirements.txt file specifying the version of jinja to use that will not cause errors in the build.