[user_accounts] Prevent crash when one user has 0 site/project affiliations (#8139)
Although this use case should never happen as it is prevented by rules in the edit_user logic, a couple bugs have let it happen in the past and the entire module becomes inaccessible. This makes the module load and displays a message in the user row informing the editor that the user does not currently have any sites and projects.

The only way to test this would be to manually delete all site associations directly in the DB user_psc_rel OR all project associations user_project_rel

Note that the module crashes even if the user not having any sites or projects is not the editor nor the editee... any user can cause this issue.