Remove pattern from email validation
This removes the regex pattern from the "email" type. It currently
breaks on addresses that have a "+" in them and is difficult to debug.
(The regex looks like it should work, but it doesn't and is very complicated.)

This only potentially impacts browsers that *do* support "pattern" but
*don't* support the "email" data type. Otherwise, the browser's built in
validation will check the validity of the email address. In the case where
neither is supported, the browser will submit and fall back to server side
validation.