[Instruments] Fix display and loading of time elements (#8160)
This fixes the display of time fields in instrument tables to be inline (hours at the left of minutes) instead of vertical (hours on top of minutes) AND fixes the loading of stored time values in the database when the time entry contains leading zeros i.e. 05:05:00

More info on time data loading: When we were using SQL data saving, SQL seems to have been fixing our time fields by adding a leading 0 to the hours or minutes that were <10 so even when the code was sending 5:5:00, SQL was converting it to 05:05:00. Since we started using JSONdata instruments, this correction has not been happening and any time value saved without correction CAN NOT be reloaded when the instrument is reloaded thus magically erasing data.