[imaging browser] add download buttons for BIDS JSON file (and NIfTI bval, bvec files for DWI) (#8159)
Currently, the image download button says 'Download MINC' which is wrong when the file attached is a NIfTI file or another image file format. Instead, the "Download MINC" has been rephrased to a more generic option "Download Image".

For DWI NIfTI files, users should also be able to download the BVAL and BVEC files that come with NIfTI DWI files. This PR adds a download button for each of those files when they are linked to the image file in the parameter_file table.

Similarly, for BIDS datasets, there are JSON files that comes with the image file that should be made available for download so this PR adds a download button to the imaging browser if a JSON file is linked to the NIfTI image file in the parameter_file table.