Fix sizing of form elements when there is no label
when the label is empty for some elements, it still includes
a col-sm-3 element which offsets the element and no way to
suppress it. Other elements do not include the 3 column space
for the label when there is no label provided. This causes
our elements to look weird and inconsistent when used in places
that we just want the element (ie. in a flexbox row.)

This fixes the
- DateElement
- TimeElement
- NumericElement
- FileElement

So that there is no space left for the label if the label is
not provided.