[new_profile] Use API for candidate creation and swal for success (#7755)
This updates the candidate creation on the new profile page to
se the LORIS API instead of duplicating the logic in PHP. The form
already collects the exact data that a POST request to the API
requires, but submits it to a different endpoint in a form encoding
instead of a json encoding.

At the same time, the logic for the response is simplified by using
a swal instead of a (very empty looking) new page. swals are already
used for errors in the module, just not success. This change simplifies
both the code and the UX.