[JSX] Add InfoPanel component
This adds an InfoPanel component to show a message to the
user in a box that's styled to draw their attention. The
"alert alert-info" bootstrap classes weren't working for me
with our current stylesheet, so the InfoPanel wraps around what's
needed to make it look reasonable and uses a FontAwesome icon to
mark it as an information panel.

Usage:
<InfoPanel>
Message that should go in the panel, can include arbitrary other
react or html if you want.
</InfoPanel>