[Imaging Uploader] Replace 'MINC' string by 'Files' in the imaging uploader column since imaging files created and inserted are not necessarily MINC files anymore (#8205)
MINC is hardcoded in the column labels specifying the number of files created by the pipeline and the number of files inserted into the imaging browser. Since LORIS supports insertion of MINC files and NIfTI files only, it does not make sense to show 'Number of MINC created' and 'Number of MINC inserted' in the imaging uploader columns header when the files created are NIfTI files...

Replaced MINC string by Files in those columns to make it more generic to all project workflows.