[Performance] Do not load module logger from constructor (#8200)
This is an attempt to solve the speed issues with demo-24.loris.ca.

In LORIS 23, there was no LORISInstance object that was used for
getting active modules, nor was there a logger instantiated from the
constructor. The added overhead of retrieving the log settings for
every module instantiation and LorisInstance->getActiveModules
instantiating every module is adding up to a lot of overhead in
the critical path on demo-24.loris.ca.

This moves the logger instantiation to be done after the
Module::factory call in the router instead of in the constructor.
The result is that it's only done once, instead of done n times
(where n is the number of modules.)

This doesn't get demo-24.loris.ca up to the speed of demo.loris.ca,
but it's significantly closer.

It's not clear why this performance regression is only happening on
demo and not in other environments but seems to be related to
latency with database queries.