[OIDC] OpenID Connect support in LORIS.
This adds support for the OpenID Connect protocol in LORIS, allowing
us to include "Login with ...." links on the login page and delegate
to third parties which implement the protocol such as Google or Facebook
or Globus. For writing this, I used auth0 (because they allow http://localhost
URLs for development.)

When a user clicks the "Login with ...." link on the login page, they
should be redirected to the third party to authenticate the request. If
they authorize it, LORIS will check if a user exists with the email address
returned. If it already exists, they will be logged in. If not, they will
be redirected to the request account page with the known information (name
and email address) populated. (We can not determine their site or project
or radiologist/examiner status from the third party.)