[login] Make it possible to set default page and values via query params (#8251)
This makes it possible to link to the Request Account page by making it
possible to set a default mode in the login module via the page=?? query
param. It also makes it possible to set the firstname, lastname, or email
address by passing them as query params.

This is useful for authentication workflows (such as OpenID Connect) that
might want to direct the user to sign up for an account and pre-populate
it with known values.