[imaging browser] Download NIfTI file from imaging browser when a NIfTI file is available in addition to a MINC file (#8206)
This modifies the imaging browser so that NIfTI files associated to a MINC file can be downloaded through the imaging browser view session page. Currently, we can download BVAL and BVEC files when they are present for DWI but could not download the NIfTI file that comes with it.

This is a bug fix for projects using MINC files and producing NIfTI files at the time of insertion so that they can download.