[tools] Fix wrong path in recreate_conflicts.php
This fixes the wrong path in the require_once for recreate_conflict.
It also passes a missing LorisInstance object to the ConflictDetector
function.

I still wasn't able to run the script to completion because I don't
have instruments that are properly configured on my sandbox for this
version of LORIS, but this resolves #8063.

Someone with up to date instruments should test the rest of the script.